output/e-commerce_platform/fullstack/fs1/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
httpx==0.25.2
pytest-cov==4.1.0
email-validator==2.1.0
